Move all types, as well as store and ledger logic, to src/
